+++ src/sf-as-network.cpp
@@ -165,7 +165,7 @@ Rcpp::List rcpp_sf_as_network (const Rcpp::List &sf_lines,
 //' @param graph Rcpp::DataFrame containing the graph
 //' @param pts Rcpp::DataFrame containing the routing points
 //'
-//' @return Rcpp::NumericVector index into graph of nearest points
+//' @return 0-indexed Rcpp::NumericVector index into graph of nearest points
 //'
 //' @noRd
 // [[Rcpp::export]]